<refactoring> : 다리 길이 입력 기능 개선 입력 값에 대한 검증 메서드에서 길이, 이동 방향, 게임 재시작 여부를 각각 검증할 수 있도록 개선함. 또, 각각에 대한 검증을 위해 Enum 타입을 별도로 선언하고 패키지를 구분지어 빼두었음.
