+++ src/main/java/bridge/Application.java
@@ -1,5 +1,7 @@
 package bridge;
 
+import bridge.UI.Input.InputView;
+
 public class Application {
     public static void main(String[] args) {
         InputView inputView = new InputView();
