+++ src/main/java/bridge/UI/Input/ValidateType.java
@@ -1,0 +1,5 @@
+package bridge.UI.Input;
+
+public enum ValidateType {
+    SIZE, MOVE, COMMAND
+}
