+++ src/main/java/bridge/UI/Output/OutputView.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.UI.Output;
 
 /**
  * 사용자에게 게임 진행 상황과 결과를 출력하는 역할을 한다.
